--- client/components/Footer_Contact_Info.jsx
@@ -1,19 +1,0 @@
-import React from 'react';
-
-export default function FooterContactInfo() {
-    return (
-        <div class = 'footer_contact_box'>
-            <div className='footer_image'>
-                <img src='./images/LOGO.jpeg' />
-            </div>
-            <ul>
-                <li>
-                    <address> Address: REDACTED ADDRESS </address>
-                </li>
-                <li>
-                    <phone> Phone: (RER) EDAC - TED? </phone>
-                </li>
-            </ul>
-        </div>
-    );
-}
