+++ client/footer.css
@@ -21,16 +21,16 @@
 .FooterSiteNav {
     @apply flex justify-between;
     margin: 20px;
-}
-
-ul {
-    list-style: none; 
-    padding: 0; 
-}
-
-li {
-    @apply flex justify-center;
-    @apply mx-2;
+    
+    ul {
+        list-style: none; 
+        padding: 0; 
+    }
+    
+    li {
+        @apply flex justify-center;
+        @apply mx-2;
+    }
 }
 
 .FooterHeader {
